Consider scc-SA mappings which map to all namespace SAs
The scc "users" element includes mappings for service accounts
in two different forms:
  system:serviceaccount:<namespace>
  system:serviceaccount:<namespace>:<serviceaccount>

Without this commit, the plugin only works correctly for the second
type. It fails for the first because it assumes that the fourth
element will always be present if the third is there. In the
first case, the scc would map to *all* service accounts in the
namespace, so this commit checks for the fourth element and handles
it properly when it is not found.

(cherry picked from commit 7d6f2d689139e73f8c4a904108092b9e7e2e7b97)