For imagestreamtags referencing ImageStreamImage, copy by tag
For ImageStreamTags, when the tag references something external
(docker image, ImageStreamTag, or ImageStreamImage), we were
copying any internal images without tag and relying on the
ImageStreamTag restore to recreate the tags. This exposes us to
an upstream Velero issue where a prior ImageStreamTag restore
(the one which should have created the tag or image which this
one references) completes but the resource isn't yet available,
so the ImageStreamTag resource fails to restore.

This commit modifies the behavior for local-namespace ImageStreamImage
references -- copying these by tag into the registry rather than
restoring via the ImageStreamTag resource. The race condition still
potentially exists for references to other namespaces included in the backup
or references by ImageStreamTag.